Fix bug in --copyall due to case insensitivity.
The matching_video function checks for videos with both lowercase and
uppercase extensions. However, at least on my machine (WSL),
os.path.exists returns true regardless of case. This works fine for
converting the video. However, the --copyall logic is case sensitive, so
you end up with a motion photo and a video when you should just have a
motion photo.

The fix is to make matching_video return the file path in its original
casing. As far as I can tell, the only way to do this is to first query
the directory for its file names and then look for a match.
os.path.realpath() and pathlib.Path.resolve() both seem to return the
path in whatever casing you queried it with. I'm using glob here.